refactor(renderer): move _new_design navigation to app/navigation/
Third slice of dissolving _new_design/. applet-sidebar-button and
group-area-sidebar move to app/navigation/ — the start of the app/ shell
folder. The remaining elements/navigation/ components join them in a
later step (see plans/elements-reorganization.md, step 4).

Pure file moves + import-path updates; no behaviour change.
